refactor: move tableDisplay
